Move StampedFile into its own file
